+++ components/nimbus/android/src/test/java/org/mozilla/experiments/nimbus/NimbusTests.kt
@@ -19,6 +19,7 @@ import mozilla.telemetry.glean.config.Configuration
 import mozilla.telemetry.glean.net.HttpStatus
 import mozilla.telemetry.glean.net.PingUploader
 import mozilla.telemetry.glean.testing.GleanTestRule
+import org.json.JSONException
 import org.json.JSONObject
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
@@ -71,6 +72,7 @@ class NimbusTests {
     private fun createNimbus(
         coenrollingFeatureIds: List<String> = listOf(),
         recordedContext: RecordedContext? = null,
+        block: Nimbus.() -> Unit = {},
     ) = Nimbus(
         context = context,
         appInfo = appInfo,
@@ -80,7 +82,7 @@ class NimbusTests {
         observer = null,
         delegate = nimbusDelegate,
         recordedContext = recordedContext,
-    )
+    ).also(block)
 
     @get:Rule
     val gleanRule = GleanTestRule(context)
@@ -734,21 +736,49 @@ class NimbusTests {
         assertEquals("Event count must match", isReadyEvents.count(), 3)
     }
 
-    @Test
-    fun `Nimbus records context if it's passed in`() {
-        class TestRecordedContext : RecordedContext {
-            var recordCount = 0
+    class TestRecordedContext(
+        private var eventQueries: MutableMap<String, Any>? = null,
+    ) : RecordedContext {
+        var recorded = mutableListOf<JSONObject>()
 
-            override fun record() {
-                recordCount++
+        override fun getEventQueries(): JsonObject {
+            val queriesJson = JSONObject()
+            for ((key, value) in eventQueries ?: mapOf()) {
+                queriesJson.put(key, value)
             }
+            return queriesJson
+        }
 
-            override fun toJson(): JsonObject {
-                val contextToRecord = JSONObject()
-                contextToRecord.put("enabled", true)
-                return contextToRecord
+        override fun setEventQueryValues(json: JsonObject) {
+            for (key in json.keys()) {
+                try {
+                    eventQueries?.put(key, json.getDouble(key))
+                } catch (exception: JSONException) {
+                    continue
+                }
             }
         }
+
+        override fun record() {
+            recorded.add(this.toJson())
+        }
+
+        override fun toJson(): JsonObject {
+            val contextToRecord = JSONObject()
+            contextToRecord.put("enabled", true)
+            val queries = this.getEventQueries()
+            for (key in queries.keys()) {
+                if (queries.get(key)::class == String::class) {
+                    queries.remove(key)
+                }
+            }
+            contextToRecord.put("events", queries)
+            return contextToRecord
+        }
+    }
+
+    @Test
+    fun `Nimbus records context if it's passed in`() {
         val context = TestRecordedContext()
         val nimbus = createNimbus(recordedContext = context)
 
@@ -761,7 +791,31 @@ class NimbusTests {
             job.join()
         }
 
-        assertEquals(context.recordCount, 1)
+        assertEquals(context.recorded.size, 1)
+    }
+
+    @Test
+    fun `Nimbus recorded context event queries are run and the value is written back into the object`() {
+        val context = TestRecordedContext(
+            mutableMapOf(
+                "TEST_QUERY" to "'event'|eventSum('Days', 1, 0)",
+            ),
+        )
+        val nimbus = createNimbus(recordedContext = context) {
+            this.recordEvent("event")
+        }
+
+        suspend fun getString(): String {
+            return testExperimentsJsonString(appInfo, packageName)
+        }
+
+        val job = nimbus.applyLocalExperiments(::getString)
+        runBlocking {
+            job.join()
+        }
+
+        assertEquals(context.recorded.size, 1)
+        assertEquals(context.recorded[0].getJSONObject("events").getDouble("TEST_QUERY"), 1.0, 0.0)
     }
 }