+++ components/nimbus/src/stateful/behavior.rs
@@ -7,6 +7,7 @@ use crate::{
     stateful::persistence::{Database, StoreId},
 };
 use chrono::{DateTime, Datelike, Duration, TimeZone, Utc};
+use regex::Regex;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use std::collections::vec_deque::Iter;
@@ -62,6 +63,7 @@ impl PartialEq for Interval {
         self.to_string() == other.to_string()
     }
 }
+
 impl Eq for Interval {}
 
 impl Hash for Interval {
@@ -84,7 +86,7 @@ impl FromStr for Interval {
             _ => {
                 return Err(NimbusError::BehaviorError(
                     BehaviorError::IntervalParseError(input.to_string()),
-                ))
+                ));
             }
         })
     }
@@ -365,7 +367,7 @@ impl EventQueryType {
                 return Err(NimbusError::TransformParameterError(format!(
                     "event transform {} requires a positive number as the second parameter",
                     self
-                )))
+                )));
             }
         } as usize;
         let zero = &Value::from(0);
@@ -375,7 +377,7 @@ impl EventQueryType {
                 return Err(NimbusError::TransformParameterError(format!(
                     "event transform {} requires a positive number as the third parameter",
                     self
-                )))
+                )));
             }
         } as usize;
 
@@ -402,7 +404,7 @@ impl EventQueryType {
                 return Err(NimbusError::TransformParameterError(format!(
                     "event transform {} requires a positive number as the second parameter",
                     self
-                )))
+                )));
             }
         } as usize;
 
@@ -425,6 +427,13 @@ impl EventQueryType {
             | Self::AveragePerNonZeroInterval => self.validate_counting_arguments(args)?,
             Self::LastSeen => self.validate_last_seen_arguments(args)?,
         })
+    }
+
+    pub fn validate_query(maybe_query: &str) -> Result<bool> {
+        let regex = Regex::new(
+            r#"^["'][^"']+["']\|event(?:Sum|LastSeen|CountNonZero|Average|AveragePerNonZeroInterval)\(["'](?:Years|Months|Weeks|Days|Hours|Minutes)["'], \d+(?:, \d+)?\)$"#,
+        )?;
+        Ok(regex.is_match(maybe_query))
     }
 
     fn error_value(&self) -> f64 {