+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # v130.0 (In progress)
 
+## ⚠️ Breaking Changes ⚠️
+
+### Nimbus SDK ⛅️🔬🔭
+- Added methods to `RecordedContext` for retrieving event queries and setting their values back to the foreign object ([#6322](https://github.com/mozilla/application-services/pull/6322)).
+
 ### Suggest
 - Added support for Fakespot suggestions.